Main.py & StreamTransformer.py draft
StreamTransformer replaces _ApiListener as a generic class for
transforming stream data.

It also includes FHCTStreamTransformer which is a specific version of
StreamTransformer.

Main.py is where the program starts
