+++ .gitignore
@@ -1,0 +1,4 @@
+
+src/run_terminal.command
+
+src/run.command
