--- docs/data_out.txt
@@ -1,6 +1,0 @@
-volume,hashtag
-30,\xe3\x83\x95\xe3\x83\xa9\xe3\x83\x8f\xe3\x83\x86\xe3\x82\xa3 \xe4\xbb\x81
-23,photography
-14,news
-12,\xe3\x83\x95\xe3\x82\xa1\xe3\x83\x9f\xe3\x83\xaa\xe3\x83\xbc\xe3\x83\x91\xe3\x83\xbc\xe3\x83\x86\xe3\x82\xa3\xe3\x83\xbc
-11,news
