+++ src/DataCollector.py
@@ -1,15 +1,12 @@
 ##################################################################################
-#                          - Data Collection Classes -                           #
+#                              - Data Collection -                               #
 #                                                                                #
 #   PROGRAMMED BY: Jake Jongewaard and Jean Flaherty                             #
-#   DATE: 03-06-2017                                                             #
+#   DATE: 03-18-2017                                                             #
 #   DESCRIPTION: Handles collecting data from the Twitter Stream API and         #
 #                passing it to the DataHandler class                             #
 #                                                                                #
 #   Classes:                                                                     #
-#       _ApiListener: sub-class of tweepy StreamListener that handles incoming   #
-#                     API data                                                   #
-#                                                                                #
 #       DataCollector: implements the _ApiListener class to take data collection #
 ##################################################################################
 import datetime
@@ -17,76 +14,8 @@ import json
 import os.path
 
 from tweepy import Stream, OAuthHandler, API
-from tweepy.streaming import StreamListener
 
 from DataHandler import DataHandler
-
-
-##################################################################################
-#                           - _ApiListener -                                     #
-#                                                                                #
-#   Description: This class is used to define how to operate data on taken from  #
-#                the Twitter Stream API.  The DataHandler class is implemented   #
-#                here for data formating                                         #
-#                                                                                #
-#   Methods:                                                                     #
-#       -on_data(self, data):                                                    #
-#       -clean(self)                                                             #
-#       -write(self)                                                             #
-##################################################################################
-
-class _ApiListener(StreamListener):
-    dat_hand = DataHandler()
-    keys = ["followers_count","hashtags","created_at","text"]
-    collect_count = 20  # number of entries to collect before stopping stream
-    threshold_size = 15 # the threshold of data size where the data is trimmed
-    period = 5          # number of entries between cleaning/writing files
-    trim_size = threshold_size - period
-    entry_count = 0     # keeping track of the number of entries added
-
-    def on_data(self, data):
-        data = json.loads(data) # convert to dictionary
-        # print("data: "+str(data))
-
-        # get entry
-        entry = {}
-        entry["followers_count"] = data["user"]["followers_count"]
-        hashtags = data["entities"]["hashtags"]
-        entry["hashtags"] = [hashtag["text"] for hashtag in hashtags]
-        entry["created_at"] = data["created_at"]
-        entry["text"] = data["text"]
-
-        self.dat_hand.add(entry)
-
-        self.entry_count += 1
-        print("ADDED: ENTRY #" + str(self.entry_count))
-
-        # if we collected enough data entries
-        if self.entry_count >= self.collect_count:
-            self.clean()
-            self.write()
-            print("STREAM STOPPED")
-            return False # stops stream
-
-        # periodic clean and write
-        if self.entry_count % self.period == 0 :
-            self.clean()
-            self.write()
-
-    def clean(self):
-        # prioritize entries with more hash tags
-        # priority = lambda entry: len(entry.get("hashtags",[]))
-        # prioritize entries with more followers
-        priority = lambda entry: entry.get("followers_count",0)
-        self.dat_hand.clean(priority, self.threshold_size, self.trim_size)
-
-    def write(self):
-        date = datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S.%f")
-        abspath = os.path.abspath(os.path.dirname(__file__))
-        docspath = os.path.join(abspath, "../docs/")
-        filepath = docspath + "STREAM DATA "+ date +".txt"
-        print("WRITING TO: "+ filepath)
-        self.dat_hand.write(filepath, self.keys)
 
 ##################################################################################
 #                           - DataCollector -                                    #
@@ -154,24 +83,3 @@ class DataCollector(object):
     def stream(self, filters, streamListener):
         stream = Stream(auth=self.api.auth, listener=streamListener)
         stream.filter(track=[filters])
-
-
-if __name__ == '__main__':
-    # store api access information
-    access_token = "3911012232-MNWJkB5E5EnN8pNrqPu8TjOIJrXHs5TmsSVI6dW"
-    access_token_secret = "jjOsGntWGcxQdPZg3ZVamCdhmYjzNDOPZxMBa1zyn1Cic"
-    consumer_key = "ToqLVlPLUsqrvHokKIitbi5ps"
-    consumer_secret = "OSLHJZBOsu1nrCk476BAIDXiVCHEl7IbAwK8ZxjwMJWjOL4q3w"
-
-    collector = DataCollector(access_token, access_token_secret, consumer_key, consumer_secret)
-    collector.authenticate()
-
-    streamListener = _ApiListener()
-
-    filter = "MACHINE LEARNING"
-    # filter = "HELLO WORLD"
-    # filter = "TRUMP"
-    # filter = "PYTHON"
-
-    print("FILTER: " + filter.upper())
-    collector.stream(filter, streamListener)
